+++ CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - `protocol::TryIntoMessage` trait renamed to `TryCreateMessage` and the one trait function `try_into_message` renamed to `try_create_message` to better reflect the fact that the trait function does not consume `self`.
+- Task parameters are now separated from task struct.
+- Task callback methods `on_failure` and `on_success` are now instance methods.
+
+### Removed
+
+- `TaskContext` struct.
 
 ## [0.2.0-alpha.2] - 2019-02-24